refactor(layout): move Contributors, CTA, Features, Hero, Subject List, and Tubelight Navbar components
