+++ components/features.tsx
@@ -2,7 +2,7 @@
 import React, { useEffect, useRef } from 'react';
 import { gsap } from 'gsap';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
-import MagicBento from './Bento';
+import MagicBento from './ui/Bento';
 
 if (typeof window !== 'undefined') {
   gsap.registerPlugin(ScrollTrigger);
