+++ components/subject-list.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useRef } from 'react';
 import { gsap } from 'gsap';
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
 import { Star, Users, Download, ArrowRight, Github, ExternalLink } from 'lucide-react';
-import MaskedDiv from './masked-div';
+import MaskedDiv from './ui/masked-div';
 import Image from 'next/image';
 
 if (typeof window !== 'undefined') {
